requirements.txt
requests
msal
python-dotenv
openpyxl